ecommerce-fullstack-ai-assistant/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.9.0
python-dotenv==1.0.1
openai==1.43.0
duckduckgo-search==6.2.10
httpx==0.27.2
python-dateutil==2.9.0.post0
